+++ NEWS.md
@@ -9,6 +9,7 @@
 - add function to compare two peak list by overlap
 - add function to output csv file from list object
 - add function to perform pmd analysis
+- add function to get chemical formula
 - fix wrong object name in getmzrt/getmzrt2
 
 # enviGCMS 0.5.0